refactor: add command line parser and merge `clust.ts` with `index.ts`
